+++ app/src/main/java/woowacourse/shopping/view/cart/CartActivity.kt
@@ -14,7 +14,6 @@ import woowacourse.shopping.App
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
 import woowacourse.shopping.view.cart.adapter.CartAdapter
-import woowacourse.shopping.view.cart.vm.CartUiEvent
 import woowacourse.shopping.view.cart.vm.CartViewModel
 import woowacourse.shopping.view.cart.vm.CartViewModelFactory
 import woowacourse.shopping.view.core.ext.showToast
