+++ app/src/main/java/woowacourse/shopping/view/cart/CartUiEvent.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.view.cart.vm
+package woowacourse.shopping.view.cart
 
 sealed interface CartUiEvent {
     data class ShowCannotIncrease(val quantity: Int) : CartUiEvent
