+++ app/src/main/java/woowacourse/shopping/view/cart/adapter/CartAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemCartBinding
 import woowacourse.shopping.view.core.handler.CartQuantityHandler
-import woowacourse.shopping.view.main.vm.state.ProductState
+import woowacourse.shopping.view.main.state.ProductState
 
 class CartAdapter(
     private var items: List<ProductState>,
