+++ app/src/main/java/woowacourse/shopping/view/cart/state/CartUiState.kt
@@ -1,9 +1,9 @@
-package woowacourse.shopping.view.cart.vm
+package woowacourse.shopping.view.cart.state
 
 import woowacourse.shopping.domain.Quantity
-import woowacourse.shopping.view.main.vm.state.CartSavingState
-import woowacourse.shopping.view.main.vm.state.IncreaseState
-import woowacourse.shopping.view.main.vm.state.ProductState
+import woowacourse.shopping.view.main.state.CartSavingState
+import woowacourse.shopping.view.main.state.IncreaseState
+import woowacourse.shopping.view.main.state.ProductState
 
 data class CartUiState(
     val items: List<ProductState> = emptyList(),
