+++ app/src/main/java/woowacourse/shopping/view/main/MainUiEvent.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.view.main.vm.event
+package woowacourse.shopping.view.main
 
 sealed interface MainUiEvent {
     data class ShowCannotIncrease(val quantity: Int) : MainUiEvent
