+++ app/src/main/res/layout/item_product.xml
@@ -12,7 +12,7 @@
 
         <variable
             name="model"
-            type="woowacourse.shopping.view.main.vm.state.ProductState" />
+            type="woowacourse.shopping.view.main.state.ProductState" />
 
         <variable
             name="adapterEventHandler"
